Update refactor for round-utils
Move logic from round-utils into other services so we can remove the round-util file

Now that we have services we want to find the applicable places to put these functions instead of in a generic utils file
